+++ include/linux/usb_bam.h
@@ -320,8 +320,9 @@ int usb_bam_disconnect_pipe(u8 idx);
  * @return pipe index on success, negative value on error.
  */
 int get_bam2bam_connection_info(u8 idx,
-	u32 *usb_bam_handle, u32 *usb_bam_pipe_idx, u32 *peer_pipe_idx,
-	struct sps_mem_buffer *desc_fifo, struct sps_mem_buffer *data_fifo);
+	unsigned long *usb_bam_handle, u32 *usb_bam_pipe_idx,
+	u32 *peer_pipe_idx, struct sps_mem_buffer *desc_fifo,
+	struct sps_mem_buffer *data_fifo);
 
 /**
  * Resets the USB BAM that has A2 pipes
@@ -445,8 +446,9 @@ static inline int usb_bam_disconnect_pipe(u8 idx)
 }
 
 static inline int get_bam2bam_connection_info(u8 idx,
-	u32 *usb_bam_handle, u32 *usb_bam_pipe_idx, u32 *peer_pipe_idx,
-	struct sps_mem_buffer *desc_fifo, struct sps_mem_buffer *data_fifo)
+	unsigned long *usb_bam_handle, u32 *usb_bam_pipe_idx,
+	u32 *peer_pipe_idx, struct sps_mem_buffer *desc_fifo,
+	struct sps_mem_buffer *data_fifo)
 {
 	return -ENODEV;
 }
